refactor(ame): Rename instr/rvmatrix to instr/ame
